requirements.txt
Flask==3.0.2
Flask-RESTful==0.3.10
sqlalchemy==2.0.29
mysql-connector-python==8.3.0
python-dotenv==1.0.1
